Implement augmenting package dependencies
A unstable CLI flag is added, -Zimplicit-dep, which permits specifying a
pkgid to add to, and a package id to add via `to=from` syntax. This is
painful for users, as they must be the full package ids from `cargo
metadata` output, but this flag is intended for use in rustbuild, where
we don't particularly care.

Multiple dependencies can be added to any given package; simply repeat
the flag.